requirements.txt
fastapi
uvicorn[standard]
requests
PyJWT
cryptography
pydantic>=2
python-dotenv
semgrep